% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeligverse.R
\docType{package}
\name{zeligverse-package}
\alias{zeligverse}
\alias{zeligverse-package}
\title{zeligverse: Easily install, load, and test Zelig packages}
\description{
Provides an easy way to load Core Zelig and ancillary Zelig
packages. It also includes functionality to run all of the Zelig packages'
tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IQSS/zeligverse}
}

}
\author{
\strong{Maintainer}: Christopher Gandrud \email{zelig.zee@gmail.com}

}
\keyword{internal}
