% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeligverse_check.R
\name{zeligverse_check}
\alias{zeligverse_check}
\title{Check all zeligverse packages}
\usage{
zeligverse_check(pkgs = core, lib_dir)
}
\arguments{
\item{pkgs}{character vector of packages to test.}

\item{lib_dir}{character string with the directory where the packages are
stored. If missing, library is found with \code{\link{.Library}}. The
first library found is used.}
}
\description{
Check all zeligverse packages
}
