% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{zeligverse_deps}
\alias{zeligverse_deps}
\title{List all zeligverse dependencies}
\source{
tidyverse
}
\usage{
zeligverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
zeligverse packages.}
}
\description{
List all zeligverse dependencies
}
