% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{zeligverse_update}
\alias{zeligverse_update}
\title{Update zeligverse packages}
\source{
tidyverse
}
\usage{
zeligverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
zeligverse packages.}
}
\description{
This will check to see if all zeligverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
zeligverse_update()
}
}
