% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeligverse.R
\docType{package}
\name{zeligverse-package}
\alias{zeligverse}
\alias{zeligverse-package}
\title{zeligverse: Easily Install and Load Stable Zelig Packages}
\description{
Provides an easy way to load stable Core Zelig and ancillary Zelig
packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IQSS/zeligverse}
  \item Report bugs at \url{https://github.com/IQSS/zeligverse/issues}
}

}
\author{
\strong{Maintainer}: Christopher Gandrud \email{zelig.zee@gmail.com}

}
\keyword{internal}
